requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pydantic
python-dotenv
openai
langdetect
pypdf
